requirements.txt
# SQS E-Commerce Analytics Pipeline
# Python 3.10+

# AWS SDK
boto3==1.34.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow>=14.0.0  # Parquet support for Medallion architecture

# Database
duckdb>=1.0.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2

# Development
ipython==8.18.1
jupyter==1.0.0

# Plotting
plotly==6.5.0
